Bug fixed an debug deleted

+++ CountingSort.h
@@ -4,4 +4,4 @@
 using namespace std;
 
 void counting_sort(int [], int);
-void counting_sort_compare(int [], int, int& );
+void counting_sort_compare(int [], int, long long & );
