+++ SelectionSort.h
@@ -4,4 +4,4 @@
 using namespace std;
 
 void selection_sort(int [], int);
-void selection_sort_compare(int [], int, int& );
+void selection_sort_compare(int [], int, long long &);
