+++ heapsort.h
@@ -1,4 +1,4 @@
 #include <iostream>
 
-void heap_sort_compare(int a[], int n, int &compare);
+void heap_sort_compare(int a[], int n, long long &compare);
 void heap_sort(int a[], int n);
