+++ shakersort.cpp
@@ -32,7 +32,7 @@ void shaker_sort(int a[], int n)
     }
 }
 
-void shaker_sort_compare(int a[], int n, int &cmp_count)
+void shaker_sort_compare(int a[], int n, long long &cmp_count)
 {
     int curr = 0;
     int right = n - 1;
