+++ shakersort.h
@@ -1,3 +1,3 @@
 #include <iostream>
-void shaker_sort_compare(int a[], int n, int &cmp_count);
+void shaker_sort_compare(int a[], int n, long long &cmp_count);
 void shaker_sort(int a[], int n);
