+++ shellsort.h
@@ -2,4 +2,4 @@
 #include <iostream>
 
 void shell_sort(int array[], int n);
-void shell_sort_compare(int array[], int n, int &compare);
+void shell_sort_compare(int array[], int n, long long &compare);
